UPSTREAM:CP:ath12k:wifi: Change beacon tx power type
The driver is unable to reject user configured beacon txpower values that
are less than 0 or greater than 255.

The current data type of beacon txpower is u8. So, values under 0, gets
wrapped under, and values greater than 255, gets wrapped around ensuring
values within the range always.

The user desires to not configure beacon tx power at all for invalid
values. So, change the txpower type to s32.

The dependent change does the boundary check.

Patch-dependency: I484db1eae88e1ef7bc2b0091923c61f789650924

+++ backports-6.1-9a0dddfb3/drivers/net/wireless/ath/ath12k/core.h
@@ -628,7 +628,7 @@ struct ath12k_vap_cfg {
 	u32 he_snd_mode;
 	u32 gtx_enable;
 	u32 hwcts2self_ofdma;
-	u8 bcn_tx_power;
+	s32 bcn_tx_power;
 	u32 ampdu_aggr_size;
 	u32 amsdu_aggr_size;
 	u32 ba_bufsize;
